requirements.txt
streamlit
streamlit-sortables
python-docx
Pillow
